server/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
Babel==2.10.3
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
click==8.0.4
cryptography==37.0.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.14.0
MarkupSafe==2.1.1
mongoengine==0.24.2
passlib==1.7.4
pip==22.1.2
pkgutil_resolve_name==1.3.10
pycparser==2.21
pymongo==3.12.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
setuptools==61.2.0
six==1.16.0
speaklater==1.3
Werkzeug==2.0.3
wheel==0.37.1
WTForms==3.0.1
zipp==3.8.0